Move lottery logic into own file
